firewall-demo/function_src/requirements.txt
google-api-python-client==1.8.0
oauth2client==4.1.3
ipaddress==1.0.23
sendgrid==6.1.3
google-cloud-secret-manager==0.2.0